requirements.txt
streamlit
google-generativeai
python-dotenv
openai-whisper
openai
st-copy-to-clipboard
joblib